requirements.txt
PySide6>=6.6
matplotlib>=3.8
SQLAlchemy>=2.0
python-docx>=1.1
openpyxl>=3.1
pandas>=2.2
numpy>=1.26
docx2pdf>=0.1.8
reportlab>=4.0
pytest>=8.0
pyinstaller>=6.0
